[resrcful] Don't start blocked projects if blocker is finishing up
If a blocker is before the blocked projects in the prioritized list, we
subtract the person hours for that day and then later think that the
blocked projects are unblocked, even though they're still blocked up
until the end of that day. We could keep track of hours remaining per
project at start of day vs end of day, but for now it's simpler to just
keep track of which projects only reached finished state within the
current day loop and still treat those as open blockers.